# src/main/java/ru/myapp/dto/PaidGroupResponseDtoShort.java
package ru.myapp.dto;

public record PaidGroupResponseDtoShort(Integer id,

                                        String name,

                                        String description,

                                        Integer cost) {
}